requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
numpy==1.25.2
opencv-python==4.8.0.76
pytesseract==0.3.10
pyyaml==6.0.1
Pillow==10.0.0